misc: Serialize tilemap
Tilemap serialization is an important milestone for the development of
the prototype as it enable us to load the scene from and save it to file.
Among some important changes, we added list of tile descriptions for the
current tiles of the map as these could be different from those of the
tileset. We also added an id for the tile description to determine its
position within the tileset.