cmd: Move command
Introduce the first command of the game, the Move command, removing the
dependency of the state component from the input system. This means that
the input is no more used to change the character state, but what we do
instead is to generate move commands.

These commands are pushed into a command ring which is handled by the
entity. Move commands, in this specific case are dispatched to the state
component.

The great benefit of this design is that not only can move commands be
generated from gamepads or keyboard input, but also by things like a
network packet, a console interaction, or a spell of mind control.