cmp: Array of states
Greatly improve the character state component by storing all the state
objects in an array and keep track of the current state with a pointer.
With this approach is important to reset the states when a transition
happens.

Also, fix a bug in the destruction listener where the contact normal
should be inverted according to the order of the fixtures involved.
This is how Box2D works.